refactor: view packages
