+++ src/test/scala/ulisse/infrastructures/view/components/RectStyleTest.scala
@@ -2,7 +2,6 @@ package ulisse.infrastructures.view.components
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.must.Matchers
-import ulisse.infrastructures.view.components.Styles.*
 
 class RectStyleTest extends AnyFlatSpec with Matchers:
